Moved Refresh button logic to Menu as it's specific to container list
